fixed inset-related bug
Bug description:

This bug was causing unwanted behaviour for the h_split and v_split
functions in that they were both returning NULL if tmid was out of
range. In turn, this would cause the original polygon record in the
database to be set to NULL.

The bug was found on one of the last four splits of the polygon from the
data that comes with this repository. After being split many times, the
polygon was very thin and long.  The bug was caused by underflow/overflow
if the inset split would use an out-of-range inset
  (for example, if the corner to cut was NW, and the sqrt(target_area)
  is greater than the height of the available polygon)

Solution:

This was remedied by checking if the inset is out of range (for all cases,
NW,NE,SW,SE) and if that condition is met, directly search for a vcut
instead.